+++ internal/backends/linuxkms/renderer/skia.rs
@@ -1,13 +1,14 @@
 // Copyright © SixtyFPS GmbH <info@slint.dev>
 // SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-Royalty-free-1.1 OR LicenseRef-Slint-commercial
 
-use std::cell::Cell;
-use std::rc::{Rc, Weak};
+use std::rc::Rc;
 
-use crate::display::{Presenter, RenderingRotation};
+use crate::display::RenderingRotation;
+use crate::drmoutput::DrmOutput;
 use i_slint_core::api::PhysicalSize as PhysicalWindowSize;
 use i_slint_core::item_rendering::ItemRenderer;
 use i_slint_core::platform::PlatformError;
+use i_slint_renderer_skia::skia_safe;
 use i_slint_renderer_skia::SkiaRendererExt;
 
 pub struct SkiaRendererAdapter {
@@ -36,7 +37,7 @@ impl SkiaRendererAdapter {
                 skia_vk_surface,
             )),
             // TODO: For vulkan we don't have a page flip event handling mechanism yet, so drive it with a timer.
-            presenter: TimerBasedAnimationDriver::new(),
+            presenter: display.presenter,
             size: display.size,
         });
 
@@ -49,17 +50,20 @@ impl SkiaRendererAdapter {
     pub fn new_opengl(
         device_opener: &crate::DeviceOpener,
     ) -> Result<Box<dyn crate::fullscreenwindowadapter::FullscreenRenderer>, PlatformError> {
-        let display = crate::display::egldisplay::create_egl_display(device_opener)?;
+        let drm_output = DrmOutput::new(device_opener)?;
+        let display = crate::display::gbmdisplay::GbmDisplay::new(drm_output)?;
 
         use i_slint_renderer_skia::Surface;
         use raw_window_handle::{HasDisplayHandle, HasWindowHandle};
+
+        let (width, height) = display.drm_output.size();
+        let size = i_slint_core::api::PhysicalSize::new(width, height);
+
         let skia_gl_surface = i_slint_renderer_skia::opengl_surface::OpenGLSurface::new(
             display.window_handle().unwrap(),
             display.display_handle().unwrap(),
-            display.size,
+            size,
         )?;
-
-        let size = display.size;
 
         let renderer = Box::new(Self {
             renderer: i_slint_renderer_skia::SkiaRenderer::new_with_surface(Box::new(
@@ -74,7 +78,32 @@ impl SkiaRendererAdapter {
         Ok(renderer)
     }
 
-    pub fn new_try_vulkan_then_opengl(
+    pub fn new_software(
+        device_opener: &crate::DeviceOpener,
+    ) -> Result<Box<dyn crate::fullscreenwindowadapter::FullscreenRenderer>, PlatformError> {
+        let drm_output = DrmOutput::new(device_opener)?;
+        let display = Rc::new(crate::display::swdisplay::SoftwareBufferDisplay::new(drm_output)?);
+
+        let skia_software_surface: i_slint_renderer_skia::software_surface::SoftwareSurface =
+            DrmDumbBufferAccess { display: display.clone() }.into();
+
+        let (width, height) = display.drm_output.size();
+        let size = i_slint_core::api::PhysicalSize::new(width, height);
+
+        let renderer = Box::new(Self {
+            renderer: i_slint_renderer_skia::SkiaRenderer::new_with_surface(Box::new(
+                skia_software_surface,
+            )),
+            presenter: display,
+            size,
+        });
+
+        eprintln!("Using Skia Software renderer");
+
+        Ok(renderer)
+    }
+
+    pub fn new_try_vulkan_then_opengl_then_software(
         device_opener: &crate::DeviceOpener,
     ) -> Result<Box<dyn crate::fullscreenwindowadapter::FullscreenRenderer>, PlatformError> {
         #[allow(unused_assignments)]
@@ -88,6 +117,10 @@ impl SkiaRendererAdapter {
         #[cfg(feature = "renderer-skia-opengl")]
         if result.is_err() {
             result = Self::new_opengl(device_opener);
+        }
+
+        if result.is_err() {
+            result = Self::new_software(device_opener);
         }
 
         result
@@ -131,59 +164,30 @@ impl crate::fullscreenwindowadapter::FullscreenRenderer for SkiaRendererAdapter 
         self.presenter.clone().register_page_flip_handler(event_loop_handle)
     }
 }
-
-struct TimerBasedAnimationDriver {
-    timer: i_slint_core::timers::Timer,
-    next_animation_frame_callback: Cell<Option<Box<dyn FnOnce()>>>,
+struct DrmDumbBufferAccess {
+    display: Rc<crate::display::swdisplay::SoftwareBufferDisplay>,
 }
 
-impl TimerBasedAnimationDriver {
-    fn new() -> Rc<Self> {
-        Rc::new_cyclic(|self_weak: &Weak<Self>| {
-            let self_weak = self_weak.clone();
-            let timer = i_slint_core::timers::Timer::default();
-            timer.start(
-                i_slint_core::timers::TimerMode::Repeated,
-                std::time::Duration::from_millis(16),
-                move || {
-                    let Some(this) = self_weak.upgrade() else { return };
-                    // Stop the timer and let the callback decide if we need to continue. It will set
-                    // `needs_redraw` to true of animations should continue, render() will be called,
-                    // present_with_next_frame_callback() will be called and then the timer restarted.
-                    this.timer.stop();
-                    if let Some(next_animation_frame_callback) =
-                        this.next_animation_frame_callback.take()
-                    {
-                        next_animation_frame_callback();
-                    }
-                },
-            );
-            // Activate it only when we present a frame.
-            timer.stop();
-
-            Self { timer, next_animation_frame_callback: Default::default() }
-        })
-    }
-}
-
-impl Presenter for TimerBasedAnimationDriver {
-    fn is_ready_to_present(&self) -> bool {
-        true
-    }
-
-    fn register_page_flip_handler(
-        self: Rc<Self>,
-        _event_loop_handle: crate::calloop_backend::EventLoopHandle,
-    ) -> Result<(), PlatformError> {
-        Ok(())
-    }
-
-    fn present_with_next_frame_callback(
+impl i_slint_renderer_skia::software_surface::RenderBuffer for DrmDumbBufferAccess {
+    fn with_buffer(
         &self,
-        ready_for_next_animation_frame: Box<dyn FnOnce()>,
-    ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-        self.next_animation_frame_callback.set(Some(ready_for_next_animation_frame));
-        self.timer.restart();
-        Ok(())
+        size: PhysicalWindowSize,
+        render_callback: &mut dyn FnMut(
+            std::num::NonZeroU32,
+            std::num::NonZeroU32,
+            skia_safe::ColorType,
+            &mut [u8],
+        )
+            -> Result<(), i_slint_core::platform::PlatformError>,
+    ) -> Result<(), i_slint_core::platform::PlatformError> {
+        let Some((width, height)) = size.width.try_into().ok().zip(size.height.try_into().ok())
+        else {
+            // Nothing to render
+            return Ok(());
+        };
+
+        self.display.map_back_buffer(&mut |mut pixels| {
+            render_callback(width, height, skia_safe::ColorType::BGRA8888, pixels.as_mut())
+        })
     }
 }
